balancing.py: Minor change; featureSelection.py: Minor change; motive.py: minor change; model.py: minor change

+++ scripts/model/model.py
@@ -19,8 +19,9 @@ from nltk.corpus import stopwords
 from nltk.tokenize import word_tokenize
 import nltk
 
-##TODO Add to the evaluation function something that creates a dataframe where all necessary data for evaluation (accuracy, f1, recall, ...) is stored for easier overview
 class Model:
+    """The Model class provides every function necessary for loading the prepared data, training and evaluating the model and finally predicting
+    """    
 
     my_tags = ['belonging', 'meaning', 'efficacy', 'distinctivness', 'self esteem', 'continuity']
 
@@ -29,70 +30,157 @@ class Model:
         self.y = None
 
     def __import_X_train(self, npz_filepath):
+        """Imports the X train data.
+
+        :param npz_filepath: Provide the filepath including "/NPZs/"
+        :type npz_filepath: string
+        :return: X sparse matrix for training
+        :rtype: sparse matrix
+        """        
         X = sc.sparse.load_npz(npz_filepath)
         return X
     
     def __import_X_test(self, npz_filepath):
+        """Imports the X test data.
+
+        :param npz_filepath: Provide the filepath including "/NPZs/"
+        :type npz_filepath: string
+        :return: X sparse matrix for testing
+        :rtype: sparse matrix
+        """       
         X_test = sc.sparse.load_npz(npz_filepath)
         return X_test
     
     def __import_Y_train(self, database, collection):
+        """Imports the Y train data.
+
+        :param database: The database where the data is stored
+        :type database: string
+        :param collection: The collection where the data is stored
+        :type collection: string
+        :return: Y data frame for training
+        :rtype: Pandas data frame
+        """        
         y = mc.getCollection(database, collection)
         y = y.drop(["_id", "id"], axis=1)
         return y
 
     def __import_Y_test(self, database, collection):
+        """Imports the Y test data.
+
+        :param database: The database where the data is stored
+        :type database: string
+        :param collection: The collection where the data is stored
+        :type collection: string
+        :return: Y data frame for testing
+        :rtype: Pandas data frame
+        """        
         y_test = mc.getCollection(database, collection)
         y_test = y_test.drop(["_id", "id"], axis=1)
         return y_test
     
     def import_train_test_data(self, filepath, database, balancing_technique, feature_selection, fs_function):
+        """Runs the functions __import_X_train, __import_X_test, __import_Y_train and __import_Y_test regarding the given input.
+
+        :param filepath: Provide the filepath including "/NPZs/"
+        :type filepath: string
+        :param database: The database where the data is stored
+        :type database: string
+        :param balancing_technique: The name of the balancing technique ("SMOTEENN", "NearMiss", "SMOTETomek","SMOTE", "TomekLinks")
+        :type balancing_technique: string
+        :param feature_selection: If the used data should use any feature selection data
+        :type feature_selection: bool
+        :param fs_function: The name of the feature selection function ("None", "chi2", "f_classif")
+        :type fs_function: string
+        :return: The data for training and testing the data
+        :rtype: Sparse matrix, sparse matrix, Pandas data frame, Pandas data frame
+        """        
         if feature_selection == True:
             X = self.__import_X_train(npz_filepath=filepath + balancing_technique + "_x_matrix_fs_" + fs_function + ".npz")
+
+            x_test_fp = filepath + "X_test.npz"
+            X_test = self.__import_X_test(x_test_fp)
+
+            fs = pi.load(open(filepath + '/Feature_' + balancing_technique + 'fs_' + fs_function + '.tchq', 'rb'))
+            X_test = fs.transform(X_test)
         else:
             X = self.__import_X_train(npz_filepath=filepath + balancing_technique + "_x_matrix.npz")
+            x_test_fp = filepath + "X_test.npz"
+            X_test = self.__import_X_test(x_test_fp)
 
-        x_test_fp = filepath + "X_test.npz"
-        X_test = self.__import_X_test(x_test_fp)
         y = self.__import_Y_train(database=database, collection=balancing_technique + "_y")
         y_test = self.__import_Y_test(database=database, collection="y_test")
         return X, X_test, y, y_test
 
     def train_model(self, model, X, y):
+        """Function for training the model
+
+        :param model: Model class object
+        :type model: class object
+        :param X: The X train data
+        :type X: Sparse matrix
+        :param y: The y train data
+        :type y: Pandas data frame
+        :return: The trained model
+        :rtype: Model object
+        """        
         trained_model = model.fit(X, y)
         return trained_model
     
     def save_model(self, model, filepath):
+        """Function for saving the model to a specific filepath
+
+        :param model: The model object
+        :type model: Model object
+        :param filepath: The filepath where the model should be stored including the model name
+        :type filepath: string
+        :return: Print statement that the model was saved
+        :rtype: string
+        """        
         pi.dump(model, open(filepath, 'wb'))
         return print("Model saved")
 
     def load_model(self, filepath):
+        """Function for loading a model
+
+        :param filepath: The filepath where the model is stored including the model name
+        :type filepath: string
+        :return: The loaded model object
+        :rtype: Model object
+        """        
         model = pi.load(open(filepath))
         return model
 
-    def load_features(self, filepath, balancing, balancing_tech, fs_tech, X_test):
-        if balancing==True:
-            filepath = filepath + balancing_tech + fs_tech + ".tchq"
-            fs = pi.load(open(filepath, 'rb'))
-            x_pred = fs.transform(X_test)
-            return x_pred
-        else:
-            filepath = filepath + fs_tech + ".tchq"
-            fs = pi.load(open(filepath, 'rb'))
-            x_pred = fs.transform(X_test)
-            return x_pred
-
     def evaluate(self, model, x_pred, y_test, average, normalize_cm, save_cm, filepath):
+        """Function for evaluating a model
+
+        :param model: The model object which should be evaluated
+        :type model: Model object
+        :param x_pred: X test data
+        :type x_pred: Sparse matrix
+        :param y_test: Y test data
+        :type y_test: Pandas data frame
+        :param average: Which average function should be used
+        :type average: string
+        :param normalize_cm: If the confusion matrix should be normalized ("true", "false")
+        :type normalize_cm: string
+        :param save_cm: If the confusion matrix should be stored
+        :type save_cm: bool
+        :param filepath: The filepath where the confusion matrix should be stored including the filename
+        :type filepath: string
+        :return: accuracy, f1, precision, recall
+        :rtype: float, float, float, float
+        """        
 
         y_pred = model.predict(x_pred)
         y_true = y_test
 
-        print('classification_report\n %s' % classification_report(y_true=y_true, y_pred=y_pred, target_names=self.my_tags))
-        print('accuracy\n %s' % accuracy_score(y_true=y_true, y_pred=y_pred))
-        print('f1-score\n %s' % f1_score(y_true=y_true, y_pred=y_pred, average=average))
-        print('precision_score\n %s' % precision_score(y_true=y_true, y_pred=y_pred, average=average))
-        print('recall_score\n %s' % recall_score(y_true=y_true, y_pred=y_pred, average=average))
-        #print('confusion matrix\n %s' % confusion_matrix(y_true, y_pred, labels=self.my_tags))
+        # print('classification_report\n %s' % classification_report(y_true=y_true, y_pred=y_pred, target_names=self.my_tags))
+        # print('accuracy\n %s' % accuracy_score(y_true=y_true, y_pred=y_pred))
+        # print('f1-score\n %s' % f1_score(y_true=y_true, y_pred=y_pred, average=average))
+        # print('precision_score\n %s' % precision_score(y_true=y_true, y_pred=y_pred, average=average))
+        # print('recall_score\n %s' % recall_score(y_true=y_true, y_pred=y_pred, average=average))
+        # print('confusion matrix\n %s' % confusion_matrix(y_true, y_pred, labels=self.my_tags))
         
         accuracy = accuracy_score(y_true=y_true, y_pred=y_pred)
         f1 = f1_score(y_true=y_true, y_pred=y_pred, average=average)
@@ -109,6 +197,15 @@ class Model:
         return accuracy, f1, precision, recall
 
     def predict_model(self, filepath_model, balancing_techniques):
+        """Function for predicting stuff right away on the console
+
+        :param filepath_model: The filepath the model directory
+        :type filepath_model: string
+        :param balancing_techniques: Which balancing techniques should be included
+        :type balancing_techniques: list
+        :return: [description]
+        :rtype: [type]
+        """        
         vocab_collection = mc.getCollection(db="09_TrainingData", col="CountVectorVocabulary")
         vocab_list = vocab_collection['dict'].to_list()
         vocab = yaml.safe_load(vocab_list[0])
@@ -249,21 +346,21 @@ filepath_Model = "D:/OneDrive - SRH IT/06 Case Study I/02 Input_Data/03 Model/Mo
 filepath_Eval = "D:/OneDrive - SRH IT/06 Case Study I/02 Input_Data/03 Model/Model_Eval_Test/"
 
 # Create empty evaluation dataframe
-eval_frame = pd.DataFrame(columns=["model", "Balancing", "Features", "Accuracy", "F1-Score", "Precision", "Recall"])
+eval_frame = pd.DataFrame(columns=["Model", "Balancing", "Features", "Accuracy", "F1-Score", "Precision", "Recall"])
 
 # Model input parameters
 seed = 69
-n_cores = -1
+n_cores = 4
 iterations = 1000 # Logistic Regression, SVM: Higher value means longer running
-estimators = 10 # Random Forest: Higher value means longer running
+estimators = 50 # Random Forest: Higher value means way longer running
 alpha = 0.1
 learning_rate = "optimal"
 
 # Istanciate classes
 modeller = Model()
+nb = MultinomialNB()
 dt = DecisionTreeClassifier(random_state=seed)
 logreg = LogisticRegression(n_jobs=n_cores, max_iter=iterations)
-nb = MultinomialNB()
 rf = RandomForestClassifier(n_jobs=n_cores, n_estimators=estimators, random_state=seed)
 svm = SGDClassifier(  # New parameters added -> experimental
     n_jobs=n_cores,
@@ -276,21 +373,34 @@ svm = SGDClassifier(  # New parameters added -> experimental
     tol=None)
 
 # Model selection parameters
-models = [dt, logreg, nb, rf, svm]
+# models = [nb, dt, logreg, rf, svm]
+# balancingTechniques = ["SMOTEENN", "NearMiss", "SMOTETomek","SMOTE", "TomekLinks"]
+# featureSelections = ["None", "chi2", "f_classif"]
+
+models = [rf]
 balancingTechniques = ["SMOTEENN", "NearMiss", "SMOTETomek","SMOTE", "TomekLinks"]
 featureSelections = ["None", "chi2", "f_classif"]
+
+total_models = len(models) * len(balancingTechniques) * len(featureSelections)
 
 for model in models:
     for balancingTechnique in balancingTechniques:
         for featureSelection in featureSelections:
-            print("Model: " + str(model) + ", Balancing: " + balancingTechnique + ", Features: " + featureSelection + " started!")
+            modelname = str(model)
+            
+            print("Model: " + modelname + ", Balancing: " + balancingTechnique + ", Features: " + featureSelection + " started!")
             # import train and test data
+            if featureSelection == "None":
+                feature_selection = False
+            else:
+                feature_selection = True
+
             X, X_test, y, y_test = modeller.import_train_test_data(
                 filepath=filepath_NPZ,
                 database="09_TrainingData",
                 balancing_technique=balancingTechnique,
-                feature_selection=False,
-                fs_function=featureSelections)
+                feature_selection=feature_selection,
+                fs_function=featureSelection)
 
             trained_model = modeller.train_model(model=model, X=X, y=y)
 
@@ -301,10 +411,10 @@ for model in models:
                 average="macro",
                 normalize_cm="true",
                 save_cm=True,
-                filepath=filepath_Eval + str(model) + "_" + balancingTechnique + "_" + featureSelection + ".png")
+                filepath=filepath_Eval + modelname + "_" + balancingTechnique + "_" + featureSelection + ".png")
             
             temp_d = {
-                "Model": str(model),
+                "Model": modelname,
                 "Balancing": balancingTechnique,
                 "Features": featureSelection,
                 "Accuracy": accuracy,
@@ -317,14 +427,24 @@ for model in models:
 
             modeller.save_model(
                 model=trained_model,
-                filepath=filepath_Model + str(model) + "_" + balancingTechnique + "_" + featureSelection + ".model")
+                filepath=filepath_Model + modelname + "_" + balancingTechnique + "_" + featureSelection + ".model")
 
-            print("Model: " + str(model) + ", Balancing: " + balancingTechnique + ", Features: " + featureSelection + " done!")
+            print("Model: " + modelname + ", Balancing: " + balancingTechnique + ", Features: " + featureSelection + " done!")
+            total_models = total_models - 1
+            print(str(total_models) + " models left.")
 
 print("Modelling done")
 
-eval_frame.reset_index(inplace=True)
-eval_frame.to_csv(filepath_Eval + "Eval_Overview.csv")
+##TODO check if files exists, if yes load and then append reset index and then save, no reset index and save
+
+if isfile(filepath_Eval + "Eval_Overview.csv"):
+    existing_eval_frame = pd.read_csv(filepath_Eval + "Eval_Overview.csv")
+    eval_frame = pd.concat([existing_eval_frame, eval_frame])
+    eval_frame.reset_index(inplace=True)
+    eval_frame.to_csv(filepath_Eval + "Eval_Overview.csv")
+else:
+    eval_frame.reset_index(inplace=True)
+    eval_frame.to_csv(filepath_Eval + "Eval_Overview.csv")
 
 # modeller.predict_model(
 #     filepath_model=filepath_Model,
